requirements.txt
Flask
PyPDF2
python-docx
scikit-learn
nltk
spacy
numpy
pandas
pdfminer.six